+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file.  
 This project follows [Semantic Versioning](https://semver.org/).
 
-## [0.9.2-b.5]
+## [0.9.2-b.5] 2025-11-03
 ### Added
 - "Reject Sweep" button
 - Batch Sweep extraction